+++ packages/core/src/modules/EventManager/GestureController.ts
@@ -1,4 +1,12 @@
-import { onClickCallback, onHoverCallback, onZoomCallback } from "../../types";
+import {
+    onClickCallback,
+    onRightClickCallback,
+    onHoverCallback,
+    onMouseDownCallback,
+    onMouseLeaveCallback,
+    onMouseUpCallback,
+    onZoomCallback,
+} from "../../types";
 import { ICamera } from "../Camera";
 import { Config } from "../Config";
 import { CoordinateTransformer } from "../CoordinateTransformer";
@@ -19,10 +27,11 @@ export class GestureController {
     private lastPinchCenter = { x: 0, y: 0 };
 
     public onClick?: onClickCallback;
+    public onRightClick?: onRightClickCallback;
     public onHover?: onHoverCallback;
-    public onMouseDown?: () => void;
-    public onMouseUp?: () => void;
-    public onMouseLeave?: () => void;
+    public onMouseDown?: onMouseDownCallback;
+    public onMouseUp?: onMouseUpCallback;
+    public onMouseLeave?: onMouseLeaveCallback;
     public onZoom?: onZoomCallback;
 
     constructor(
@@ -34,6 +43,35 @@ export class GestureController {
         private onCameraChange: () => void
     ) {}
 
+    private getEventCoords = (e: MouseEvent | { clientX: number; clientY: number }) => {
+        const rect = this.canvas.getBoundingClientRect();
+        const mouseX = e.clientX - rect.left;
+        const mouseY = e.clientY - rect.top;
+        const world = this.transformer.screenToWorld(mouseX, mouseY);
+        const screen = this.transformer.worldToScreen(Math.floor(world.x), Math.floor(world.y));
+
+        return {
+            coords: {
+                raw: world,
+                snapped: { x: Math.floor(world.x), y: Math.floor(world.y) },
+            },
+            mouse: {
+                raw: { x: mouseX, y: mouseY },
+                snapped: {
+                    x: screen.x,
+                    y: screen.y,
+                },
+            },
+            client: {
+                raw: { x: e.clientX, y: e.clientY },
+                snapped: {
+                    x: screen.x + rect.left,
+                    y: screen.y + rect.top,
+                },
+            },
+        };
+    };
+
     handleClick = (e: MouseEvent) => {
         if (this.shouldPreventClick) {
             this.shouldPreventClick = false;
@@ -42,37 +80,24 @@ export class GestureController {
         if (!this.config.get().eventHandlers.click || !this.onClick) {
             return;
         }
-        const rect = this.canvas.getBoundingClientRect();
-        const mouseX = e.clientX - rect.left;
-        const mouseY = e.clientY - rect.top;
-        const world = this.transformer.screenToWorld(mouseX, mouseY);
-        const screen = this.transformer.worldToScreen(Math.floor(world.x), Math.floor(world.y));
+        const { coords, mouse, client } = this.getEventCoords(e);
+        this.onClick(coords, mouse, client);
+    };
 
-        this.onClick(
-            {
-                raw: world,
-                snapped: { x: Math.floor(world.x), y: Math.floor(world.y) },
-            },
-            {
-                raw: { x: e.clientX - rect.left, y: e.clientY - rect.top },
-                snapped: {
-                    x: screen.x,
-                    y: screen.y,
-                },
-            },
-            {
-                raw: { x: e.clientX, y: e.clientY },
-                snapped: {
-                    x: screen.x + rect.left,
-                    y: screen.y + rect.top,
-                },
-            }
-        );
+    handleContextMenu = (e: MouseEvent) => {
+        if (!this.config.get().eventHandlers.rightClick || !this.onRightClick) {
+            return;
+        }
+        e.preventDefault();
+        const { coords, mouse, client } = this.getEventCoords(e);
+
+        this.onRightClick(coords, mouse, client);
     };
 
     handleMouseDown = (e: MouseEvent) => {
         if (this.onMouseDown) {
-            this.onMouseDown();
+            const { coords, mouse, client } = this.getEventCoords(e);
+            this.onMouseDown(coords, mouse, client);
         }
 
         if (!this.config.get().eventHandlers.drag) {
@@ -87,29 +112,9 @@ export class GestureController {
     handleMouseMove = (e: MouseEvent) => {
         if (!this.isDragging) {
             if (this.onHover && this.config.get().eventHandlers.hover) {
-                const rect = this.canvas.getBoundingClientRect();
-                const mouseX = e.clientX - rect.left;
-                const mouseY = e.clientY - rect.top;
-                const world = this.transformer.screenToWorld(mouseX, mouseY);
-                const screen = this.transformer.worldToScreen(Math.floor(world.x), Math.floor(world.y));
+                const { coords, mouse, client } = this.getEventCoords(e);
 
-                this.onHover(
-                    { raw: world, snapped: { x: Math.floor(world.x), y: Math.floor(world.y) } },
-                    {
-                        raw: { x: e.clientX - rect.left, y: e.clientY - rect.top },
-                        snapped: {
-                            x: screen.x,
-                            y: screen.y,
-                        },
-                    },
-                    {
-                        raw: { x: e.clientX, y: e.clientY },
-                        snapped: {
-                            x: screen.x + rect.left,
-                            y: screen.y + rect.top,
-                        },
-                    }
-                );
+                this.onHover(coords, mouse, client);
             }
             return;
         }
@@ -125,18 +130,20 @@ export class GestureController {
         this.onCameraChange();
     };
 
-    handleMouseUp = () => {
+    handleMouseUp = (e: MouseEvent) => {
         if (this.onMouseUp) {
-            this.onMouseUp();
+            const { coords, mouse, client } = this.getEventCoords(e);
+            this.onMouseUp(coords, mouse, client);
         }
 
         this.isDragging = false;
         this.canvas.style.cursor = this.config.get().cursor.default || "default";
     };
 
-    handleMouseLeave = () => {
+    handleMouseLeave = (e: MouseEvent) => {
         if (this.onMouseLeave) {
-            this.onMouseLeave();
+            const { coords, mouse, client } = this.getEventCoords(e);
+            this.onMouseLeave(coords, mouse, client);
         }
 
         this.isDragging = false;