+++ packages/core/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `@canvas-tile-engine/core` will be documented in this file.
 
+## [0.0.3] - 2025-12-21
+
+* feat: unify mouse event callback structures by @enesyukselx in https://github.com/enesyukselx/canvas-tile-engine/pull/24
+* feat: add onRightClick callback for right-click events by @enesyukselx in https://github.com/enesyukselx/canvas-tile-engine/pull/30
+* fix: ensure initial canvas top/left are set to 0 by @enesyukselx in https://github.com/enesyukselx/canvas-tile-engine/pull/25
+* feat: add gridAligned config option for pixel-perfect grid alignment by @enesyukselx in https://github.com/enesyukselx/canvas-tile-engine/pull/28
+* feat: add getVisibleBounds() to get visible viewport coordinates by @enesyukselx in https://github.com/enesyukselx/canvas-tile-engine/pull/29
+
 ## [0.0.2] - 2025-12-18
 
 ### Added